requirements.txt
absl-py==2.1.0
accelerate==0.30.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.2
aiohttp==3.10.7
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.6.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
av==14.0.1
beautifulsoup4==4.12.3
bitarray @ file:///home/conda/feedstock_root/build_artifacts/bitarray_1748068052211/work
bitsandbytes==0.41.0
black==24.1.0
blinker==1.8.2
boto3==1.35.29
botocore==1.35.86
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
cmake==3.31.2
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
Cython==3.1.2
DataProperty==1.1.0
datasets==2.16.1
decorator==5.1.1
decord==0.6.0
deepspeed==0.13.5
diffusers @ git+https://github.com/huggingface/diffusers.git@6a5f06488c0d88c1827c016835cd5f64abe4b52c
dill==0.3.7
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
environs==11.0.0
et-xmlfile==1.1.0
eval_type_backport==0.2.2
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.2.0
-e git+https://github.com/pytorch/fairseq@ecbf110e1eb43861214b05fa001eff584954f65a#egg=fairseq
fastapi==0.115.0
ffmpeg==1.4
ffmpy==0.4.0
filelock==3.16.1
flash-attn==2.3.6
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.3.1
future==1.0.0
gdown==5.2.0
gensim==4.3.3
gguf==0.10.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1745509363867/work
gradio==3.35.2
gradio_client==0.2.9
grpcio==1.66.1
h11==0.14.0
h5py==3.12.1
hf_transfer==0.1.9
hjson==3.1.0
httpcore==0.16.3
httptools==0.6.4
httpx==0.23.3
huggingface-hub==0.25.1
humanfriendly==10.0
humanize==4.10.0
hydra-core==1.0.7
identify==2.6.6
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
interegular==0.3.3
ipython==8.18.1
isort==5.13.2
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lark==1.2.2
latex2mathml==3.77.0
lazy_loader==0.4
librosa==0.10.2.post1
linkify-it-py==2.0.3
lit==18.1.8
llvmlite==0.43.0
# Editable Git install with no remote (lmms_eval==0.2.3)
-e /mnt/hwfile/renyiming/ly_workspace/video_bench/aurora/src/lmms-eval
loguru==0.7.3
lxml==5.3.0
Markdown==3.7
markdown-it-py==2.2.0
markdown2==2.5.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
mdit-py-plugins==0.3.3
mdurl==0.1.2
MiniSom==2.3.3
missingno==0.5.2
mmcls==0.25.0
mmcv-full==1.6.2
mmsegmentation==0.30.0
model-index==0.1.11
modelscope==1.22.3
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.15
multiprocessing-logging==0.3.4
mypy-extensions==1.0.0
narwhals==1.8.4
nest-asyncio==1.6.0
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1698504735452/work
ninja==1.11.1.1
nltk==3.9.1
nodeenv==1.9.1
numba==0.60.0
numexpr==2.10.2
numpy==1.26.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.0.6
openai==1.51.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.5
openxlab==0.0.11
optree @ file:///home/conda/feedstock_root/build_artifacts/optree_1748442456222/work
ordered-set==4.1.0
orjson==3.10.7
outlines==0.0.46
packaging==24.1
pandas==2.2.3
parso==0.8.4
partial-json-parser==0.2.1.1.post4
pathspec==0.12.1
pathvalidate==3.2.3
peft==0.13.0
petrel-oss-sdk===v2.2.1-2-g1505ef3-master
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
portalocker==3.1.1
pre_commit==4.1.0
prettytable==3.11.0
prometheus_client==0.21.0
prompt_toolkit==3.0.50
protobuf==3.20.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==17.0.0
pyarrow-hotfix==0.6
pybind11 @ file:///D:/bld/pybind11-split_1747934313674/work
pybind11_global @ file:///home/conda/feedstock_root/build_artifacts/pybind11-split_1747934270866/work
pycocoevalcap==1.2
pycocotools==2.0.8
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pynvml==11.5.3
pyparsing==3.1.4
PySocks==1.7.1
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.10
pytz==2024.2
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1737454647378/work
pyzmq==26.2.0
referencing==0.35.1
regex @ file:///home/conda/feedstock_root/build_artifacts/regex_1730952178579/work
requests==2.32.3
rfc3986==1.5.0
rich==13.8.1
rpds-py==0.20.0
s3transfer==0.10.2
sacrebleu @ file:///home/conda/feedstock_root/build_artifacts/sacrebleu_1735995758027/work
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.11.1
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==3.3.1
sentencepiece==0.2.0
sentry-sdk==2.20.0
setproctitle==1.3.4
sgl-kernel==0.0.3.post1
sglang==0.3.6
shortuuid==1.0.13
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
soundfile==0.13.0
soupsieve==2.6
soxr==0.5.0.post1
sqlitedict==2.1.0
stack-data==0.6.3
starlette==0.38.6
streamlit==1.38.0
streamlit-image-select==0.6.0
svgwrite==1.4.3
sympy==1.13.1
tabledata==1.3.4
tabulate @ file:///home/conda/feedstock_root/build_artifacts/tabulate_1733589744265/work
tcolorpy==0.1.7
tenacity==8.3.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==0.9.12
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
torch==2.0.0
torchao==0.8.0
torchaudio==2.0.1
TorchCRF==1.1.0
torchvision==0.15.2
tornado==6.4.1
tqdm==4.66.5
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transformers==4.37.2
transformers-stream-generator==0.0.5
typepy==1.3.4
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==1.26.20
uvicorn==0.31.0
uvloop==0.21.0
virtualenv==20.29.1
wandb==0.19.4
watchdog==4.0.2
watchfiles==0.24.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.4
wrapt==1.16.0
xgboost==2.1.2
xxhash==3.5.0
yacs==0.1.8
yapf==0.40.1
yarl==1.13.1
yt-dlp==2025.1.15
zipp==3.20.2
zss==1.2.0
zstandard==0.23.0
